fix: grpcio_tools not installed
At a first glance this seems like the corresponding pip package is not
installed, however this is not the case. To check for the protoc compiler
in the CMake file,`python3 -m grpc_tools.protoc' is called and
'Missing input file' is expected as the output. With any other output
the build will fail.

Due to a deprecation warning in pkg_resources Python package the output is different
from the expected, although the package is installed. There as been
a fix from Google, however it does only apply for Python verions >= 3.9
https://github.com/grpc/grpc/commit/eae9c558cf75ec6eada593361ae03e9d60dc7011

The implemented solution is to split the string into a list (after each \n) and
get the first element to retrieve the expected output. This will fix the
build, but the deprecation warnings are still printed